L-CAST/inference-service/requirements.txt
fastapi
uvicorn
python-dotenv
requests
numpy
scikit-learn
pandas
sentence-transformers